requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
psycopg2-binary==2.9.10
pgvector==0.3.6
openai==1.59.5
anthropic==0.42.0
pydantic==2.10.4
tiktoken==0.8.0
numpy==2.2.1
sentence-transformers==3.4.1
scikit-learn==1.6.1
httpx==0.28.1
pdfplumber>=0.11
python-multipart>=0.0.9
google-generativeai==0.8.3
